+++ R/ci_wald.R
@@ -38,6 +38,7 @@ ci_wald <- function(model, ci = .95, dof = NULL, effects = c("fixed", "random", 
       "kenward" = ,
       "kr" = se_kenward(model),
       "ml1" = se_ml1(model),
+      "betwithin" = se_betwithin(model),
       "satterthwaite" = se_satterthwaite(model),
       standard_error(model, component = component)
     )