+++ R/p_value_betwithin.R
@@ -1,6 +1,6 @@
-#' "m-l-1" approximation for SEs, CIs and p-values
+#' Between-within approximation for SEs, CIs and p-values
 #'
-#' Approximation of degrees of freedom based on a "m-l-1" heuristic as suggested by Elff et al. (2019).
+#' Approximation of degrees of freedom based on a "between-within" heuristic.
 #'
 #' @param model A mixed model.
 #' @param dof Degrees of Freedom.
@@ -12,34 +12,30 @@
 #' is small (even if the sample size of level-1 units is high). In such cases
 #' it is recommended to approximate a more accurate number of degrees of freedom
 #' for such inferential statitics (see \cite{Li and Redden 2015}). The
-#' \emph{m-l-1} heuristic is such an approach that uses a t-distribution with
-#' fewer degrees of freedom (\code{dof_ml1}) to calculate p-values
-#' (\code{p_value_ml1}), standard errors (\code{se_ml1}) and confidence intervals
-#' (\code{ci(method = "ml1")}).
+#' \emph{Between-within} denominator degrees of freedom approximation is
+#' recommended in particular for generalized linear mixed models with repeated
+#' measurements (longitudinal design). \code{dof_betwithin}) implements a heuristic
+#' based on the between-within approach. \strong{Note} that this implementation
+#' does not return exactly the same results as shown in \cite{Li and Redden 2015},
+#' but similar.
 #' }
 #' \subsection{Degrees of Freedom for Longitudinal Designs (Repeated Measures)}{
 #' In particular for repeated measure designs (longitudinal data analysis),
-#' the \emph{m-l-1} heuristic is likely to be more accurate than simply using the
-#' residual or infinite degrees of freedom, because \code{dof_ml1()} returns
-#' different degrees of freedom for within-cluster and between-cluster effects.
+#' the \emph{between-within} heuristic is likely to be more accurate than simply
+#' using the residual or infinite degrees of freedom, because \code{dof_betwithin()}
+#' returns different degrees of freedom for within-cluster and between-cluster effects.
 #' }
-#' \subsection{Limitations of the "m-l-1" Heuristic}{
-#' Note that the "m-l-1" heuristic is not applicable (or at least less accurate)
-#' for complex multilevel designs, e.g. with cross-classified clusters. In such cases,
-#' more accurate approaches like the Kenward-Roger approximation (\code{dof_kenward()})
-#' is recommended. However, the "m-l-1" heuristic also applies to generalized
-#' mixed models, while approaches like Kenward-Roger or Satterthwaite are limited
-#' to linear mixed models only.
-#' }
-#' @seealso \code{dof_ml1()} and \code{se_ml1()} are small helper-functions
+#' @seealso \code{dof_betwithin()} and \code{se_betwithin()} are small helper-functions
 #' to calculate approximated degrees of freedom and standard errors of model
-#' parameters, based on the "m-l-1" heuristic.
+#' parameters, based on the "between-within" heuristic.
 #'
 #' @examples
 #' \donttest{
 #' if (require("lme4")) {
-#'   model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
-#'   p_value_ml1(model)
+#'   data(sleepstudy)
+#'   model <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)
+#'   dof_betwithin(model)
+#'   p_value_betwithin(model)
 #' }
 #' }
 #' @return A data frame.
@@ -49,16 +45,16 @@
 #' }
 #' @importFrom stats pt coef
 #' @export
-p_value_ml1 <- function(model, dof = NULL) {
-  UseMethod("p_value_ml1")
+p_value_betwithin <- function(model, dof = NULL) {
+  UseMethod("p_value_betwithin")
 }
 
 
 
 #' @export
-p_value_ml1.merMod <- function(model, dof = NULL) {
+p_value_betwithin.default <- function(model, dof = NULL) {
   if (is.null(dof)) {
-    dof <- dof_ml1(model)
+    dof <- dof_betwithin(model)
   }
   .p_value_dof(model, dof)
 }