+++ R/standard_error.R
@@ -317,6 +317,8 @@ standard_error.merMod <- function(model, effects = c("fixed", "random"), method 
         # ml1 approx
       } else if (method == "ml1") {
         se_ml1(model)
+      } else if (method == "betwithin") {
+        se_betwithin(model)
         # Kenward approx
       } else if (method %in% c("kenward", "kr")) {
         se_kenward(model)